ingestion/requirements.txt
kafka-python
requests
praw
qdrant-client
sentence-transformers
dotenv
sys 
os
uuid
httpx